Added separate ast_node file. Implemented scoping for more expressions. Added some scoping unit tests
