+++ src/bin.rs
@@ -4,6 +4,7 @@ use std::io::prelude::*;
 extern crate grace_lib;
 use grace_lib::parser;
 use grace_lib::utils::*;
+use grace_lib::ast_node::*;
 use grace_lib::bytecode::*;
